requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==5.0
django-cors-headers==4.3.1
django-elasticsearch-dsl==8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.11.0
elasticsearch==8.11.1
elasticsearch-dsl==8.11.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.18
